refactor : runner 인터페이스 변경 및 cliRunner로 분리
