+++ src/main/java/org/example/article/ArticleApplication.java
@@ -3,6 +3,7 @@ package org.example.article;
 import org.example.article.cli.DispatchController;
 import org.example.article.cli.config.CommandConfig;
 import org.example.article.cli.config.DefaultCommandConfig;
+import org.example.article.cli.runner.CliRunner;
 
 public class ArticleApplication {
 
@@ -14,7 +15,7 @@ public class ArticleApplication {
 
     public static void main(String[] args) {
         DispatchController dispatchController = new DispatchController(commandConfig.commandController());
-        ApplicationRunner applicationRunner = new ApplicationRunner(dispatchController);
+        ApplicationRunner applicationRunner = new CliRunner(dispatchController);
         applicationRunner.run();
     }
 }
