+++ src/main/java/org/example/article/cli/runner/CliRunner.java
@@ -1,14 +1,13 @@
-package org.example.article;
+package org.example.article.cli.runner;
 
-import org.example.article.cli.ApplicationStateHolder;
 import org.example.article.cli.DispatchController;
 import org.example.article.cli.view.InputView;
 
-public class ApplicationRunner {
+public class CliRunner implements org.example.article.ApplicationRunner {
 
     private final DispatchController dispatchController;
 
-    public ApplicationRunner(DispatchController dispatchController) {
+    public CliRunner(DispatchController dispatchController) {
         this.dispatchController = dispatchController;
     }
